requirements.txt
scapy
netaddr
pycrypto